new utility file and fixed partial level gamble bug
