+++ utils/BasicUtility.py
@@ -23,40 +23,5 @@ def custom_command_list():
 def clean(i):
   return i.lower().replace(' ','')
 
-def calculate_xp_for_lvl(level):
-  return int(sum(i+300*pow(2, i/7) for i in range(1,level+1))/4)
 
-def generate_level_up_image(level, url):
-  from PIL import Image
-  from PIL import ImageFont
-  from PIL import ImageDraw
-  import io
-  import requests
 
-  # Image stuff
-  th_size = 50, 50
-  img = Image.open("./data/levelup/levelup_bg2.jpg")
-  fd = requests.get(url)
-  i_f = io.BytesIO(fd.content)
-  avatar = Image.open(i_f)
-  draw = ImageDraw.Draw(img, 'RGBA')
-
-  # Font stuff
-  font = "./data/Roboto-Black.ttf"
-  name_font = ImageFont.truetype(font, 12)
-  level_font = ImageFont.truetype(font, 26)
-  base_color = (60, 60, 70)
-  exp_color = (190, 190, 200)
-
-  # # Avatar background filler
-  draw.rectangle([13,8,67,62], fill=(0,0,0,50))
-
-  # # Avatar handling
-  a_im = avatar.crop((0,0,128,128)).resize(th_size, Image.ANTIALIAS)
-  img.paste(a_im, (15,10,65,60))
-
-  # Text implementation
-  draw.text((15, 63), "Level Up!", base_color, font=name_font)
-  draw.text((25 if level > 9 else 33,72), str(level), base_color, font=level_font)
-
-  img.save('./data/levelup/level_up.jpg')
